+++ src/components/common/ExhibitionCarousel.tsx
@@ -2,7 +2,7 @@ import { Card, CardContent } from '../ui/card';
 import { Carousel, CarouselContent, CarouselItem, CarouselNext, CarouselPrevious } from '../ui/carousel';
 import Autoplay from 'embla-carousel-autoplay';
 
-export function LiveCarousel() {
+export function ExhibitionCarousel() {
   return (
     <div className='col-span-3 flex justify-center'>
       <Carousel className='w-2/3' plugins={[Autoplay({ delay: 4000 })]}>
